Quick fix to issue #24 (reported upstreams)
Ideally, I might also fix this in makeTerm. The quick fix replaces
missing obo_ids when parsing the query resonse

  if (anyNA(resp$obo_id)) {
      i <- is.na(resp$obo_id)
      resp$obo_id[i] <- sub("_", ":", resp$short_form)[i]
  }

but when the terms are created, obo_id is still missing

> (qtrms <- as(qry, "Terms"))
Object of class 'Terms' with 7 entries
 From  7 ontologies
  NULL, OMIT:0020822 ... GO:0005802, GO:0005802
> qtrms[[1]]
A Term from the NCIT ontology:
 Label: Trans-Golgi Network
  A network of membrane components where vesicles bud off the Golgi
  apparatus to bring proteins, membranes and other components to the
  plasma membrane, secretory vesicles and endosomes.
> qtrms[1]
Object of class 'Terms' with 1 entries
 From the NCIT ontology
NULL

This should probably be addressed in makeTerm, but let's wait and see
what the OLS has to say (see https://github.com/EBISPOT/OLS/issues/153).